feat: icon 이동 및 footer 제거
