+++ src/App.tsx
@@ -7,16 +7,18 @@ export const App: React.FC = () => {
   return (
     <BrowserRouter>
       <Compt.Header />
-      <S.Content>
+      <S.Wrapper>
         <Compt.LeftSide />
-        <Routes>
-          <Route path={PAGE_URL.WELCOME} element={<Compt.Login />} />
-          <Route path={PAGE_URL.FRIENDS} element={<div />} />
-          <Route path={PAGE_URL.CHAT} element={<div />} />
-          <Route path={PAGE_URL.SETTING} element={<Compt.Canvas />} />
-          <Route path="*" element={<div>Wrong Path</div>} />
-        </Routes>
-      </S.Content>
+        <S.MainSide>
+          <Routes>
+            <Route path={PAGE_URL.WELCOME} element={<Compt.Login />} />
+            <Route path={PAGE_URL.FRIENDS} element={<Compt.Friends />} />
+            <Route path={PAGE_URL.CHAT} element={<div />} />
+            <Route path={PAGE_URL.DRAWING} element={<Compt.Canvas />} />
+            <Route path="*" element={<div>Wrong Path</div>} />
+          </Routes>
+        </S.MainSide>
+      </S.Wrapper>
     </BrowserRouter>
   );
 };
