--- src/components/Footer/index.ts
@@ -1,1 +1,0 @@
-export { FooterContainer as Footer } from './FooterContainer';
