+++ src/configs/url.ts
@@ -2,12 +2,12 @@ export const ROUTE_URL = {
   WELCOME: '/',
   FRIENDS: 'friends',
   CHAT: 'chat',
-  SETTING: 'setting',
+  DRAWING: 'drawing',
 };
 
 export const PAGE_URL = {
   WELCOME: '/',
   FRIENDS: `/${ROUTE_URL.FRIENDS}`,
   CHAT: `/${ROUTE_URL.CHAT}`,
-  SETTING: `/${ROUTE_URL.SETTING}`,
+  DRAWING: `/${ROUTE_URL.DRAWING}`,
 };
