+++ src/styled.tsx
@@ -2,12 +2,12 @@ import styled from 'styled-components';
 
 export const Wrapper = styled.div`
   display: flex;
-  with: 100%;
-  height: 100%;
-`;
-
-export const Content = styled.div`
   width: 100%;
   height: calc(100% - 50px);
   padding-top: 50px;
+`;
+
+export const MainSide = styled.div`
+  width: calc(100% - 60px);
+  height: 100%;
 `;
